Slight update to Scanner
Test cases failed:
AndEmptyString(): No lexical exception was thrown. I suppose the final \* should result in a lexical error because there is only one quotation mark inside the string. However, if it is \\*, then it should pass. \* fails but \\* passes. Right now, it is just stuck running.
illegalLineTermInStringLiteral() and andIllegalCarriageReturn(): Both have an issue of reading \ inside the string. It treats pos 1 as an invalid token. I am pretty sure these cases were caught in the catch statement. I tried fixing this issue all night and I was unable to find a solution for it. If my Scanner class was not corrupted, it would give a failure but not an error.
stringContainBSlash2() and stringContainBSlash4(): \\* I believe were supposed to end the " but they failed